+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 # Yii VarDumper
 
 
-## 1.0.6 under development
+## 1.1.0 under development
 
-- no changes in this release.
-
+- Enh #54: Add support for exporting objects with closures (yiiliveext)
 
 ## 1.0.5 March 17, 2021